renamed lib, working on test
this commit renames the org lib and continues to add core functionality
that will make the initial test pass. so far the program can read the
grammar file, and save all of the references into a context together
with the value of that reference.

so far the program can only do this for objects and when the reference
refers to a string.

the next big feature to develop is the logic that handles the
replacement of the reference with the actual data.