working test
this commit adds the final piece to the puzzle of how to replace the
references with actual value. as mentioned in the previous commit msg,
this is only for the mvp test.

the next step will be to add more tests and to make the solution better
in terms of width, and alignment with the underlying json-specification.