added superset functionality
this commit adds some additional features to the origional json spec.
the data parser will be able to parse not only double quotes, but also
single quotes and backticks.

i also made it possible to leave dangling commas at the end of
key-pairs, making it easier to generate data sets, as you dont have to
include logic that checks if the last inserted element was the last!