align with json spec
this commit takes one step back and aligns the spec with the json spec.
i started to wotk on a json test, making sure that all of the different
types are parsed correctly. as of thios commit, the program can parse:
- null
- number
- string
- object (ive only tested one level)

i need to add more tests for especially escaped characters and unicodes.
i also found out that the references needs to be parsed so we can
reference a number and place it into a string for example