improved json compatibility
this commit improves the json compatibility and the json test for serde
and this lib now output the same result! to make this possible i had to
add string parsing for the object key as well.

i added a check so that references cant be added into the key, i should
probably add a test for this as well!