refactor folder structure
this commit renames the previous parser mod to ast, since i think this
is a better description of what the pest parser produces. then i renamed
the main module to parser as it's translating the ast into json and
applying the additional rules that ive created, making it a super
parser!