added support for references within arrays
this commit adds the abillity to write references from within objects
that are inside of an array. the solution right now is far from optimal
though, as the program will loop over every element in the array, and
see if that element contains the right key.

the idea that ive got for improving this logic is to trace the reference
in the firstplace and save the index into the path.