changed ref identifier to & instead of $
this commit replaces the reference identifyer to an & since i think it's
a better symbol for what is happening.

this commit also removes $ and _ as valid inital symbols for keys in
objects. i think it's bad practice to use this symbols to indicate meta
information, rather you should use something more expressive as a meta
key...